agent_results/end2end_codex/iclr2025_dl4c/codex/requirements.txt
datasets
transformers
torch
faiss-cpu
pandas
matplotlib
sacrebleu
sentencepiece
protobuf
